cleanup: low-hanging fruit
