--- libs/Math/include/math/geo/_circle_decl.h
@@ -1,19 +1,0 @@
-#pragma once
-
-#include <sized.h>
-
-#include <math/vector.h>
-
-namespace math {
-using namespace sized; // NOLINT(*-using-namespace)
-
-namespace geo {
-
-struct Circle {
-	Vec3 center;
-	Vec3 normal;
-	flt radius;
-};
-
-} // namespace geo
-} // namespace math
