--- libs/Math/include/math/geo/_circle_impl.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "math/geo/_circle_impl.hpp"
-
-namespace math::geo {
-
-} // namespace math::geo
