--- libs/Math/include/math/geo/_ray_decl.h
@@ -1,22 +1,0 @@
-#pragma once
-
-#include <sized.h>
-
-#include "math/vector.h"
-
-namespace math {
-using namespace sized; // NOLINT(*-using-namespace)
-
-namespace geo {
-
-struct Ray {
-	Vec3 origin;
-	Vec3 delta;
-
-	constexpr Ray() = default;
-	constexpr Ray(const Vec3& origin, const Vec3& delta);
-	constexpr Ray(const Vec3& origin, const Vec3& direction, flt length);
-};
-
-} // namespace geo
-} // namespace math
