--- libs/Math/include/math/geo/_sphere_impl.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#include "math/geo/_sphere_decl.h"
-
-namespace math::geo {
-
-
-
-} // namespace math::geo
