+++ libs/Math/include/math/geo/ray.h
@@ -1,8 +1,30 @@
 #pragma once
 
-#include "math/geo/_ray_decl.h"
-#include "math/geo/_ray_impl.hpp"
+#include <sized.h>
 
-namespace math::geo {
-struct Ray;
-}
+#include "math/vector.h"
+
+namespace math {
+using namespace sized; // NOLINT(*-using-namespace)
+
+namespace geo {
+
+struct Ray {
+	Vec3 origin;
+	Vec3 delta;
+
+	constexpr Ray() = default;
+
+	constexpr Ray(const Vec3& origin, const Vec3& delta)
+		: origin(origin)
+		, delta(delta)
+	{}
+
+	constexpr Ray(const Vec3& origin, const Vec3& direction, flt length)
+		: origin(origin)
+		, delta(direction * length)
+	{}
+};
+
+} // namespace geo
+} // namespace math
